结构调整，删除无效资源
