+++ Sources/HXPhotoPicker/Picker/Extension/Picker+PHAssetCollection.swift
@@ -1,5 +1,5 @@
 //
-//  Core+PHAssetCollection.swift
+//  Picke+PHAssetCollection.swift
 //  HXPhotoPicker
 //
 //  Created by Slience on 2021/1/7.
