+++ Sources/HXPhotoPicker/Picker/Model/LivePhotoError.swift
