+++ Sources/HXPhotoPicker/Picker/Util/AssetManager+AVAsset.swift
