+++ Sources/HXPhotoPicker/Picker/Util/AssetManager+AVAssetExportSession.swift
