+++ Sources/HXPhotoPicker/Picker/Util/AssetManager+Asset.swift
