+++ Sources/HXPhotoPicker/Picker/Util/AssetManager+Image.swift
