+++ Sources/HXPhotoPicker/Picker/Util/AssetManager+ImageData.swift
