+++ Sources/HXPhotoPicker/Picker/Util/AssetManager+LivePhoto.swift
