+++ Sources/HXPhotoPicker/Picker/Util/AssetManager+LivePhotoURL.swift
